feat: separate social component
