+++ packages/island/src/theme-default/components/SocialLinks/presetIcons.tsx
@@ -1,9 +1,4 @@
-import { useState } from 'react';
-import { DefaultTheme } from 'shared/types';
-import styles from './index.module.scss';
-import type { ComponentPropsWithIsland } from 'shared/types';
-
-export const icons = {
+export default {
   discord: (
     <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
       <title>Discord</title>
@@ -149,211 +144,4 @@ export const icons = {
       ></path>
     </svg>
   )
-};
-
-interface IShownLinksProps {
-  links: DefaultTheme.SocialLink[];
-  moreIconVisible?: boolean;
-  mouseEnter: () => void;
-}
-
-type IHiddenLinksProps = Omit<
-  IShownLinksProps,
-  'moreIconVisible' | 'mouseEnter'
->;
-
-interface ILinkContentComp {
-  link: DefaultTheme.SocialLink;
-  popperStyle?: Record<string, unknown>;
-}
-
-const LinkContentComp = (props: ILinkContentComp) => {
-  const { link, popperStyle = {} } = props;
-  const { icon, mode = 'link', content } = link;
-  const IconComp = icons[icon as keyof typeof icons];
-
-  const [contentVisible, setContentVisible] = useState(false);
-  const mouseEnterIcon = () => {
-    setContentVisible(true);
-  };
-  const mouseLeavePopper = () => {
-    setContentVisible(false);
-  };
-
-  if (mode === 'link') {
-    return (
-      <a key={icon} href={content} target="_blank" rel="noopener noreferrer">
-        <div className={`${styles.socialLinksIcon}`}>{IconComp}</div>
-      </a>
-    );
-  }
-  if (mode === 'text') {
-    return (
-      <div
-        className={`${styles.socialLinksIcon}`}
-        cursor="pointer"
-        relative=""
-        onMouseEnter={mouseEnterIcon}
-        onMouseLeave={mouseLeavePopper}
-      >
-        {IconComp}
-        {contentVisible ? (
-          <div
-            absolute=""
-            z="1"
-            p="3"
-            w="50"
-            break="all"
-            pos="right-0"
-            border-1=""
-            rounded="xl"
-            bg="bg-default"
-            style={{
-              boxShadow: 'var(--island-shadow-3)',
-              ...popperStyle
-            }}
-          >
-            <div text="ml">{content}</div>
-          </div>
-        ) : null}
-      </div>
-    );
-  }
-  if (mode === 'img') {
-    return (
-      <div
-        className={`${styles.socialLinksIcon}`}
-        cursor="pointer"
-        relative=""
-        onMouseEnter={mouseEnterIcon}
-        onMouseLeave={mouseLeavePopper}
-      >
-        {IconComp}
-        {contentVisible ? (
-          <div
-            absolute=""
-            z="1"
-            p="3"
-            w="50"
-            break="all"
-            pos="right-0"
-            border-1=""
-            rounded="xl"
-            bg="bg-default"
-            style={{
-              boxShadow: 'var(--island-shadow-3)',
-              ...popperStyle
-            }}
-          >
-            <img src={content} alt="img" />
-          </div>
-        ) : null}
-      </div>
-    );
-  }
-
-  return;
-};
-
-const ShownLinksComp = (props: IShownLinksProps) => {
-  const { links, moreIconVisible = false, mouseEnter } = props;
-
-  return (
-    <div
-      h="100%"
-      flex=""
-      gap="x-2"
-      items-center=""
-      transition="color duration-300"
-    >
-      {links.map((item) => (
-        <LinkContentComp
-          key={item.icon}
-          link={item}
-          popperStyle={{ top: '2.5rem' }}
-        />
-      ))}
-      {moreIconVisible ? (
-        <div
-          className="i-carbon-chevron-sort-down"
-          onMouseEnter={mouseEnter}
-        ></div>
-      ) : null}
-    </div>
-  );
-};
-
-const HiddenLinksComp = (props: IHiddenLinksProps) => {
-  const { links } = props;
-
-  return (
-    <div
-      absolute=""
-      pos="top-13 right-0"
-      p="3"
-      w="26"
-      border-1=""
-      rounded="xl"
-      bg="bg-default"
-      style={{
-        boxShadow: 'var(--island-shadow-3)',
-        marginRight: '10px'
-      }}
-      flex="~ wrap"
-      gap="2"
-    >
-      {links.map((item) => (
-        <LinkContentComp
-          key={item.icon}
-          link={item}
-          popperStyle={{ top: '1.25rem' }}
-        />
-      ))}
-    </div>
-  );
-};
-
-export const NavSocialLinks = ({
-  socialLinks
-}: {
-  socialLinks: DefaultTheme.SocialLink[];
-} & ComponentPropsWithIsland) => {
-  const moreThanThree = socialLinks.length > 3;
-
-  const shownLinks = [];
-  const hiddenLinks = [];
-  for (const i in socialLinks) {
-    if (i < 3) {
-      shownLinks.push(socialLinks[i]);
-    } else {
-      hiddenLinks.push(socialLinks[i]);
-    }
-  }
-
-  const [moreLinksVisible, setMoreLinksVisible] = useState(false);
-  const mouseEnter = () => {
-    setMoreLinksVisible(true);
-  };
-  const mouseLeave = () => {
-    setMoreLinksVisible(false);
-  };
-
-  return (
-    <div
-      className="social-links"
-      h="100%"
-      flex=""
-      items-center=""
-      before="menu-item-before"
-      relative=""
-      onMouseLeave={mouseLeave}
-    >
-      <ShownLinksComp
-        links={shownLinks}
-        moreIconVisible={moreThanThree}
-        mouseEnter={mouseEnter}
-      />
-      {moreLinksVisible ? <HiddenLinksComp links={hiddenLinks} /> : null}
-    </div>
-  );
 };
